refactor(cookie): created new classes for meta and sameSite
